requirements.txt
Django==3.1.4
django-import-export==2.4.0
djangorestframework==3.12.2
drf-yasg==1.20.0
djangorestframework-jwt==1.11.0